refactor and fix h1
